Menambahkan Default Foto

+++ application/controllers/Barangkeluar.php
@@ -44,6 +44,7 @@ class Barangkeluar extends CI_Controller {
                     echo "Unggah file gagal!";
                 }
             } else {
+                $foto = 'default.png';
             }
             
             $dokumen = $_FILES['dokumen']['name'];
@@ -145,7 +146,9 @@ class Barangkeluar extends CI_Controller {
                 $this->load->library('upload', $config);
                 if ($this->upload->do_upload('foto')) {
                     $foto_lama = $detail->foto;
+                    if($foto_lama != 'default.png'){
                     unlink(FCPATH.'/assets/file/Barangkeluar/'.$foto_lama);
+                    }
                     $foto = $this->upload->data('file_name');
                 } else {
                     echo "Unggah file gagal!";
